Move token vendor context to contexts dir.
