+++ packages/shared-ui/src/elements/elements.ts
@@ -61,6 +61,6 @@ export { Tooltip } from "./tooltip/tooltip.js";
 export { UI } from "./ui-controller/ui-controller.js";
 export { WelcomePanel } from "./welcome-panel/welcome-panel.js";
 
-export { tokenVendorContext } from "./connection/token-vendor.js";
+export { tokenVendorContext } from "../contexts/token-vendor.js";
 export { googleDriveFileIdInputPlugin } from "./google-drive/google-drive-file-id.js";
 export { googleDriveQueryInputPlugin } from "./google-drive/google-drive-query.js";
